backend/requirements.txt
# Martin Backend Minimal Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
structlog==23.2.0
